Extract sources modal state hook
